refactor: Isolate nonce checking into a dedicated module
